Move foreign table info components into folder
