+++ src/app/admin/auditLogTable/auditLogModal/infoComponentsForForeignTables.tsx/EventName.tsx
@@ -3,9 +3,9 @@
 import React from "react";
 import supabase from "@/supabaseClient";
 import { logErrorReturnLogId } from "@/logger/logger";
-import { TextValueContainer } from "./AuditLogModal";
-import { ForeignResponse } from "./types";
-import GeneralForeignInfo from "./GeneralForeignInfo";
+import { TextValueContainer } from "../AuditLogModal";
+import { ForeignResponse } from "../types";
+import GeneralForeignInfo from "../GeneralForeignInfo";
 
 interface EventName {
     eventName: string;
